acit/requirements.txt
google-ads
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-bigquery
google-cloud-storage
six
absl-py
apache-beam
